removed BigDecimal in favor of double. TODO: Make log output pretty again

+++ src/effects/Bleed.java
@@ -1,17 +1,14 @@
 package effects;
 
-import java.math.BigDecimal;
-
 import battle.BattleSetting;
 import heroes.Hero;
 
 /**
  * Poison effect
- *
  */
 public class Bleed extends TemporaryEffect {
 
-  public Bleed(BattleSetting setting, Integer remainingRounds, Hero source, BigDecimal skillStrength) {
+  public Bleed(BattleSetting setting, Integer remainingRounds, Hero source, double skillStrength) {
     super(h -> h.receiveAttack(setting, source, skillStrength, false, false, x -> null),
         h -> h.receiveAttack(setting, source, skillStrength, false, false, x -> null), h -> null, "Poison",
         remainingRounds);
