+++ src/effects/Heal.java
@@ -1,16 +1,12 @@
 package effects;
 
-import java.math.BigDecimal;
-
 /**
  * Poison effect
- *
  */
 public class Heal extends TemporaryEffect {
 
-  public Heal(Integer remainingRounds, Integer baseStat, BigDecimal modifier) {
-    super(h -> h.heal(baseStat, modifier), h -> h.heal(baseStat, modifier), h -> null, "Heal",
-        remainingRounds);
+  public Heal(Integer remainingRounds, Integer baseStat, double modifier) {
+    super(h -> h.heal(baseStat, modifier), h -> h.heal(baseStat, modifier), h -> null, "Heal", remainingRounds);
   }
 
 }
